Ported Fudge-Java commit https://github.com/OpenGamma/Fudge-Java/commit/1cf356a495c42e47db61b5d73fa184e8f3a41580
Made the type dictionary be part of the FudgeContext.

This means that we're passing it around during encoding and decoding,
which contributed
to most of the deltas here. It also means that we're not using the
.INSTANCE very much,
so that's been made package protected (we just use it rather than
constructing too many
of them).